scripts/authentik-wireguard-sync/requirements.txt
requests>=2.31,<3
python-dotenv>=1.0,<2
cryptography>=42,<44
